Revert "Change type for va_arg to unsigned long" Change va_arg types back to unsigned int This reverts commit ee6b724f1f95591988682675ed5b03cf30d0a98c.

+++ _Task1.c
@@ -7,10 +7,10 @@
  *
  * Return: Number of character written
  */
-int print_dec(long n, BUFFER *bp)
+int print_dec(int n, BUFFER *bp)
 {
-	unsigned long divider = 1000000000;
-	unsigned long value;
+	unsigned int divider = 1000000000;
+	unsigned int value;
 	int counter = 0;
 
 	if (n < 0)
